Add GameState and keep highlighted tile

+++ src/constants.rs
@@ -1,6 +1,6 @@
 pub const SCREEN_W: f32 = 960.0;
 pub const SCREEN_H: f32 = 540.0;
 
-pub const GRID_W: isize = 16;
-pub const GRID_H: isize = 9;
+pub const GRID_W: i32 = 16;
+pub const GRID_H: i32 = 9;
 pub const GRID_THICKNESS: f32 = 4.0;
